+++ DamageAssessment/damage_assessment.py
@@ -3,10 +3,9 @@ import xarray as xr
 import pandas as pd
 from rasterio.errors import NotGeoreferencedWarning
 import warnings
-from utils.dataset import get_resolution, degrees_to_meters
-from utils.damages import apply_ddf
-from utils.geo import clip_dataarray_by_geometries
-import subprocess
+from utilities.dataset import get_resolution, degrees_to_meters
+from utilities.damages import apply_ddf
+from utilities.geo import clip_dataarray_by_geometries
 import numpy as np
 import math
