--- Geoparquet2Pmtiles/utils/gcs.py
@@ -1,82 +1,0 @@
-from google.cloud import storage
-import logging
-
-logging.basicConfig()
-logging.root.setLevel(logging.INFO)
-
-
-def download_blob(bucket_name, source_blob_name, destination_file_name):
-    """Downloads a blob from the bucket."""
-    # The ID of your GCS bucket
-    # bucket_name = "your-bucket-name"
-
-    # The ID of your GCS object
-    # source_blob_name = "storage-object-name"
-
-    # The path to which the file should be downloaded
-    # destination_file_name = "local/path/to/file"
-
-    storage_client = storage.Client()
-
-    bucket = storage_client.bucket(bucket_name)
-
-    # Construct a client side representation of a blob.
-    # Note `Bucket.blob` differs from `Bucket.get_blob` as it doesn't retrieve
-    # any content from Google Cloud Storage. As we don't need additional data,
-    # using `Bucket.blob` is preferred here.
-    blob = bucket.blob(source_blob_name)
-    blob.download_to_filename(destination_file_name)
-
-    logging.info(
-        "Downloaded storage object %s from bucket %s to local file %s.", 
-        source_blob_name, bucket_name, destination_file_name
-    )
-
-
-def upload_blob(bucket_name, source_file_name, destination_blob_name):
-    """Uploads a file to the bucket."""
-    # The ID of your GCS bucket
-    # bucket_name = "your-bucket-name"
-    # The path to your file to upload
-    # source_file_name = "local/path/to/file"
-    # The ID of your GCS object
-    # destination_blob_name = "storage-object-name"
-
-    logging.info(
-        "Uploading file %s to bucket %s as %s.",
-        source_file_name, bucket_name, destination_blob_name
-    )
-
-    storage_client = storage.Client(project='global-mangroves')
-    bucket = storage_client.bucket(bucket_name)
-    blob = bucket.blob(destination_blob_name)
-
-    # Optional: set a generation-match precondition to avoid potential race conditions
-    # and data corruptions. The request to upload is aborted if the object's
-    # generation number does not match your precondition. For a destination
-    # object that does not yet exist, set the if_generation_match precondition to 0.
-    # If the destination object already exists in your bucket, set instead a
-    # generation-match precondition using its generation number.
-    # generation_match_precondition = 0
-    logging.info("Uploading.")
-    blob.upload_from_filename(
-        source_file_name,
-        # if_generation_match=generation_match_precondition
-    )
-
-    logging.info(
-        "File %s uploaded to %s.",
-        source_file_name, destination_blob_name
-    )
-
-def list_blobs(bucket_name):
-    """Lists all the blobs in the bucket."""
-    # bucket_name = "your-bucket-name"
-
-    storage_client = storage.Client(project="global-mangroves")
-
-    # Note: Client.list_blobs requires at least package version 1.17.0.
-    blobs = storage_client.list_blobs(bucket_name)
-
-    # Note: The call returns a response only when the iterator is consumed.
-    return [blob.name for blob in blobs]
